Add predefined color for speaker label
