+++ cvat/apps/dataset_manager/bindings.py
@@ -31,9 +31,9 @@ from django.conf import settings
 from django.db.models import Prefetch, QuerySet
 from django.utils import timezone
 
-from cvat.apps.dataset_manager.formats.utils import get_label_color
 from cvat.apps.engine import models
 from cvat.apps.engine.cache import MediaCache
+from cvat.apps.engine.label_colors import get_label_color
 from cvat.apps.engine.lazy_list import LazyList
 from cvat.apps.engine.media_io.frame_provider import FrameOutputType, TaskFrameProvider
 from cvat.apps.engine.models import (
